refactoring filters system
